requirements.txt
Pillow>=9.0.0
customtkinter>=5.2.0
lxml>=4.9.0
python-docx>=0.8.11